code/pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
exclude = ["tests", "tests.*"]

[project]
name = "sqg"
version = "1.0.0"
description = "A robust and scalable alternative to existing K-means solvers."
authors = [
    { name = "Vladimir Norkin", email = "v.norkin@kpi.ua" },
    { name = "Anton Kozyriev", email = "a.kozyriev@kpi.ua" },
]
readme = "README.md"
license = { file = "LICENSE.code.md" }
dependencies = [
    "numpy>=1.26.4,<2",
    "scikit-learn>=1.5.1,<2",
    "tqdm>=4.66.0,<5",
]
requires-python = ">=3.8"
keywords = [
    "stochastic quantization",
    "clustering algorithms",
    "K-means",
    "stochastic gradient descent",
    "non-convex optimization",
]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Science/Research",
    "Intended Audience :: Education",
    "Natural Language :: English",
    "Operating System :: OS Independent",
    "License :: OSI Approved :: MIT License",
    "Topic :: Education",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
    "Topic :: Scientific/Engineering :: Mathematics",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
]

[project.urls]
Homepage = "https://github.com/kaydotdev/stochastic-quantization"
Issues = "https://github.com/kaydotdev/stochastic-quantization/issues"
Repository = "https://github.com/kaydotdev/stochastic-quantization.git"

[project.optional-dependencies]
faiss-cpu = ["faiss-cpu>=1.10.0,<2"]
faiss-gpu = ["faiss-gpu>=1.10.0,<2"]
